rename movementsensor -> movement_sensor to match casing pattern elsewhere
