+++ src/viam/sdk/components/movement_sensor/client.hpp
@@ -8,8 +8,8 @@
 #include <viam/api/component/movementsensor/v1/movementsensor.grpc.pb.h>
 
 #include <viam/sdk/common/linear_algebra.hpp>
-#include <viam/sdk/components/movementsensor/movementsensor.hpp>
-#include <viam/sdk/components/movementsensor/server.hpp>
+#include <viam/sdk/components/movement_sensor/movement_sensor.hpp>
+#include <viam/sdk/components/movement_sensor/server.hpp>
 #include <viam/sdk/config/resource.hpp>
 #include <viam/sdk/robot/client.hpp>
