+++ src/viam/sdk/components/movement_sensor/server.cpp
@@ -1,8 +1,8 @@
-#include <viam/sdk/components/movementsensor/server.hpp>
+#include <viam/sdk/components/movement_sensor/server.hpp>
 
 #include <viam/sdk/common/linear_algebra.hpp>
 #include <viam/sdk/common/utils.hpp>
-#include <viam/sdk/components/movementsensor/movementsensor.hpp>
+#include <viam/sdk/components/movement_sensor/movement_sensor.hpp>
 #include <viam/sdk/config/resource.hpp>
 #include <viam/sdk/rpc/server.hpp>
